requirements.txt
certifi==2015.4.28
defusedxml==0.4.1
Django==1.8.3
django-allauth==0.21.0
django-autoslug==1.8.0
django-bootstrap3==6.1.0
django-comments-xtd==1.4b2
django-contrib-comments==1.6.1
django-cors-headers==1.1.0
django-favit==0.2.4
django-filter==0.10.0
django-haystack==2.4.0
django-jsonify==0.3.0
django-pgjson==0.3.1
django-push-notifications==1.3.1
django-rest-auth==0.4.0
django-taggit==0.16.2
djangorestframework==3.1.3
djmail==0.10.0
elasticsearch==1.6.0
oauthlib==1.0.0
Pillow==2.9.0
psycopg2==2.6.1
pyelasticsearch==1.4
python3-openid==3.0.5
requests==2.7.0
requests-oauthlib==0.5.0
simplejson==3.8.0
six==1.9.0
sorl-thumbnail==12.3
urllib3==1.10.4
